requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.2
cssselect2==0.6.0
daphne==3.0.2
Django==3.2.13
django-braces==1.15.0
django-classy-tags==3.0.1
django-cms==3.10.0
django-countries==7.3.2
django-credit-cards==0.4.1
django-crispy-forms==1.14.0
django-formtools==2.3
django-jazzmin==2.5.0
django-model-utils==4.2.0
django-multiselectfield==0.1.12
django-sekizai==3.0.1
django-star-ratings==0.9.2
django-treebeard==4.5.1
djangocms-admin-style==3.1.0
djangocms-animate-on-scroll==1.0.1
djangorestframework==3.13.1
easy-thumbnails==2.8.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
lxml==4.8.0
mysqlclient==2.1.0
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
reportlab==3.6.9
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
swapper==1.3.0
tinycss2==1.1.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.12
webencodings==0.5.1
zope.interface==5.4.0
